Move FIPS helpers to its own file
